refactor: ram command上SF32LB52的特有实现
