--- sftool-lib/src/ram_command.rs
@@ -1,101 +1,0 @@
-use crate::SifliTool;
-use crate::sifli_debug::SifliUartCommand;
-use std::cmp::PartialEq;
-use std::io::Write;
-use strum::{Display, EnumString};
-
-#[derive(EnumString, Display, Debug, Clone, PartialEq, Eq)]
-pub enum Command {
-    #[strum(to_string = "burn_erase_all 0x{address:08x}\r")]
-    EraseAll { address: u32 },
-
-    #[strum(to_string = "burn_verify 0x{address:08x} 0x{len:08x} 0x{crc:08x}\r")]
-    Verify { address: u32, len: u32, crc: u32 },
-
-    #[strum(to_string = "burn_erase 0x{address:08x} 0x{len:08x}\r")]
-    Erase { address: u32, len: u32 },
-
-    #[strum(to_string = "burn_erase_write 0x{address:08x} 0x{len:08x}\r")]
-    WriteAndErase { address: u32, len: u32 },
-
-    #[strum(to_string = "burn_write 0x{address:08x} 0x{len:08x}\r")]
-    Write { address: u32, len: u32 },
-
-    #[strum(to_string = "burn_read 0x{address:08x} 0x{len:08x}\r")]
-    Read { address: u32, len: u32 },
-
-    #[strum(to_string = "burn_reset\r")]
-    SoftReset,
-
-    #[strum(to_string = "burn_speed {baud} {delay}\r")]
-    SetBaud { baud: u32, delay: u32 },
-}
-
-#[derive(EnumString, Display, Debug, Clone, PartialEq, Eq)]
-pub enum Response {
-    #[strum(serialize = "OK")]
-    Ok,
-    #[strum(serialize = "Fail")]
-    Fail,
-    #[strum(serialize = "RX_WAIT")]
-    RxWait,
-}
-
-pub const RESPONSE_STR_TABLE: [&str; 3] = ["OK", "Fail", "RX_WAIT"];
-
-pub trait RamCommand {
-    fn command(&mut self, cmd: Command) -> Result<Response, std::io::Error>;
-    fn send_data(&mut self, data: &[u8]) -> Result<Response, std::io::Error>;
-}
-
-// RamCommand trait is now implemented by each specific chip implementation
-// in their respective modules (sf32lb52/mod.rs, sf32lb56/mod.rs, sf32lb58/mod.rs)
-
-pub trait DownloadStub {
-    fn download_stub(&mut self) -> Result<(), std::io::Error>;
-}
-
-impl<T: SifliTool> DownloadStub for T {
-    fn download_stub(&mut self) -> Result<(), std::io::Error> {
-        self.attempt_connect()?;
-        self.download_stub_impl()?;
-
-        std::thread::sleep(std::time::Duration::from_millis(100));
-        self.port().clear(serialport::ClearBuffer::All)?;
-        self.debug_command(SifliUartCommand::Exit)?;
-
-        // 1s之内串口发b"\r\n"字符串，并等待是否有"msh >"回复，200ms发一次b"\r\n"
-        let mut buffer = Vec::new();
-        let mut now = std::time::SystemTime::now();
-        const RETRY: u32 = 5;
-        let mut retry_count = 0;
-        self.port().write_all(b"\r\n")?;
-        self.port().flush()?;
-        loop {
-            let elapsed = now.elapsed().unwrap().as_millis();
-            if elapsed > 200 {
-                retry_count += 1;
-                now = std::time::SystemTime::now();
-                self.port().write_all(b"\r\n")?;
-                self.port().flush()?;
-                buffer.clear();
-            }
-            if retry_count > RETRY {
-                return Err(std::io::Error::new(std::io::ErrorKind::TimedOut, "Timeout"));
-            }
-
-            let mut byte = [0];
-            let ret = self.port().read_exact(&mut byte);
-            if ret.is_err() {
-                continue;
-            }
-            buffer.push(byte[0]);
-
-            // 一旦buffer出现"msh >"字符串，就认为接收完毕
-            if buffer.windows(5).any(|window| window == b"msh >") {
-                break;
-            }
-        }
-        Ok(())
-    }
-}
