+++ sftool-lib/src/sf32lb58/reset.rs
@@ -1,0 +1,8 @@
+use crate::reset::Reset;
+use super::SF32LB58Tool;
+
+impl Reset for SF32LB58Tool {
+    fn soft_reset(&mut self) -> Result<(), std::io::Error> {
+        todo!("SF32LB58 soft_reset implementation not yet available")
+    }
+}
